+++ CHANGELOG.md
@@ -21,6 +21,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [0.3.2] - 2026-03-04
+
+### Added
+
+- Add filtering options to FilterBarForm for categories, location, restock needs, and status
+- Add filtering functionality to MainContainer for inventory items
+
+### Changed
+
+- Update FilterBar test to include specific filter options for Location, Needs Restock, and Status
+
+---
+
 ## [0.3.1] - 2026-03-04
 
 ### Added